Replace profile button with dropdown menu
